+++ changelog.md
@@ -16,3 +16,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
   - Updated the LICENSE file and the package.json to include the license.
   - Updated the README to include the link to the license file in the repo's wiki.
+  - Tested base counter